Split the Screen class into the SafeZone and GameScreen classes
They are now in Tools
